Update ss_cobra from TG

+++ code/modules/food&drinks/food/snacks.dm
@@ -3,16 +3,21 @@
 	desc = "yummy"
 	icon = 'icons/obj/food.dmi'
 	icon_state = null
-	var/bitesize = 1
+	var/bitesize = 2
 	var/bitecount = 0
 	var/trash = null
-	var/slice_path
+	var/slice_path    // for sliceable food. path of the item resulting from the slicing
 	var/slices_num
 	var/eatverb
 	var/wrapped = 0
 	var/dried_type = null
 	var/potency = null
 	var/dry = 0
+	var/cooked_type = null  //for microwave cooking. path of the resulting item after microwaving
+	var/filling_color = "#FFFFFF" //color to use when added to custom food.
+	var/custom_food_type = null  //for food customizing. path of the custom food to create
+	var/junkiness = 0  //for junk food. used to lower human satiety.
+
 
 	//Placeholder for effect that trigger on eating that aren't tied to reagents.
 /obj/item/weapon/reagent_containers/food/snacks/proc/On_Consume()
@@ -49,8 +54,15 @@
 		if(!canconsume(M, user))
 			return 0
 
+		var/fullness = M.nutrition + 10
+		for(var/datum/reagent/consumable/C in M.reagents.reagent_list) //we add the nutrition value of what we're currently digesting
+			fullness += C.nutriment_factor * C.volume / C.metabolization_rate
+
 		if(M == user)								//If you're eating it yourself.
-			var/fullness = M.nutrition + (M.reagents.get_reagent_amount("nutriment") * 25)
+			if(junkiness && M.satiety < -150 && M.nutrition > NUTRITION_LEVEL_STARVING + 50 )
+				M << "<span class='notice'>You don't feel like eating any more junk food at the moment.</span>"
+				return 0
+
 			if(wrapped)
 				M << "<span class='notice'>You can't eat wrapped food!</span>"
 				return 0
@@ -58,19 +70,18 @@
 				M << "<span class='notice'>You hungrily [eatverb] some of \the [src] and gobble it down!</span>"
 			else if(fullness > 50 && fullness < 150)
 				M << "<span class='notice'>You hungrily begin to [eatverb] \the [src].</span>"
-			else if(fullness > 150 && fullness < 350)
+			else if(fullness > 150 && fullness < 500)
 				M << "<span class='notice'>You [eatverb] \the [src].</span>"
-			else if(fullness > 350 && fullness < 550)
+			else if(fullness > 500 && fullness < 600)
 				M << "<span class='notice'>You unwillingly [eatverb] a bit of \the [src].</span>"
-			else if(fullness > (550 * (1 + M.overeatduration / 2000)))	// The more you eat - the more you can eat
+			else if(fullness > (600 * (1 + M.overeatduration / 2000)))	// The more you eat - the more you can eat
 				M << "<span class='notice'>You cannot force any more of \the [src] to go down your throat.</span>"
 				return 0
 		else
 			if(! (isslime(M) || isbrain(M)) )		//If you're feeding it to someone else.
-				var/fullness = M.nutrition + (M.reagents.get_reagent_amount("nutriment") * 25)
 				if(wrapped)
 					return 0
-				if(fullness <= (550 * (1 + M.overeatduration / 1000)))
+				if(fullness <= (600 * (1 + M.overeatduration / 1000)))
 					M.visible_message("<span class='danger'>[user] attempts to feed [M] [src].</span>", \
 										"<span class='userdanger'>[user] attempts to feed [M] [src].</span>")
 				else
@@ -78,7 +89,8 @@
 										"<span class='userdanger'>[user] cannot force anymore of [src] down [M]'s throat!</span>")
 					return 0
 
-				if(!do_mob(user, M)) return
+				if(!do_mob(user, M))
+					return
 				add_logs(user, M, "fed", object="[reagentlist(src)]")
 				M.visible_message("<span class='danger'>[user] forces [M] to eat [src].</span>", \
 									"<span class='userdanger'>[user] feeds [M] to eat [src].</span>")
@@ -88,6 +100,8 @@
 				return
 
 		if(reagents)								//Handle ingestion of the reagent.
+			if(M.satiety > -200)
+				M.satiety -= junkiness
 			playsound(M.loc,'sound/items/eatfood.ogg', rand(10,50), 1)
 			if(reagents.total_volume)
 				reagents.reaction(M, INGEST)
@@ -119,29 +133,31 @@
 		user << "[src] was bitten multiple times!"
 
 
-/obj/item/weapon/reagent_containers/food/snacks/attackby(obj/item/weapon/W, mob/user)
+/obj/item/weapon/reagent_containers/food/snacks/attackby(obj/item/weapon/W, mob/user, params)
 	if(istype(W,/obj/item/weapon/storage))
 		..() // -> item/attackby()
 		return 0
-	if((slices_num <= 0 || !slices_num) || !slice_path)
+	if(istype(W,/obj/item/weapon/reagent_containers/food/snacks))
+		var/obj/item/weapon/reagent_containers/food/snacks/S = W
+		if(custom_food_type && ispath(custom_food_type))
+			if(S.w_class > 2)
+				user << "<span class='warning'>The ingredient is too big for [src].</span>"
+				return 0
+			if(contents.len >= 20)
+				user << "<span class='warning'>You can't add more ingredients to [src].</span>"
+				return 0
+			var/obj/item/weapon/reagent_containers/food/snacks/customizable/C = new custom_food_type(get_turf(src))
+			C.initialize_custom_food(src, S, user)
+			return 0
+	if(is_sharp(W))
+		var/sharpness = is_sharp(W)
+		if(slice(sharpness, W, user))
+			return 1
+
+
+/obj/item/weapon/reagent_containers/food/snacks/proc/slice(var/accuracy, obj/item/weapon/W, mob/user)
+	if((slices_num <= 0 || !slices_num) || !slice_path) //is the food sliceable?
 		return 0
-	var/inaccurate = 0
-	if( \
-			istype(W, /obj/item/weapon/kitchenknife) || \
-			istype(W, /obj/item/weapon/butch) || \
-			istype(W, /obj/item/weapon/scalpel) || \
-			istype(W, /obj/item/weapon/kitchen/utensil/knife) \
-		)
-	else if( \
-			istype(W, /obj/item/weapon/circular_saw) || \
-			istype(W, /obj/item/weapon/melee/energy/sword) && W:active || \
-			istype(W, /obj/item/weapon/melee/energy/blade) || \
-			istype(W, /obj/item/weapon/shovel) || \
-			istype(W, /obj/item/weapon/hatchet) \
-		)
-		inaccurate = 1
-	else
-		return 0 // --- this is everything that is NOT a slicing implement, and which is not being slipped into food; allow afterattack ---
 
 	if ( \
 			!isturf(src.loc) || \
@@ -153,7 +169,7 @@
 		return 1
 
 	var/slices_lost = 0
-	if (!inaccurate)
+	if (accuracy > 1)
 		user.visible_message( \
 			"<span class='notice'>[user] slices [src].</span>", \
 			"<span class='notice'>You slice [src].</span>" \
@@ -164,12 +180,33 @@
 			"<span class='notice'>You inaccurately slice [src] with your [W]!</span>" \
 		)
 		slices_lost = rand(1,min(1,round(slices_num/2)))
+
+	if(!slice_path && !slices_num)
+		return
 	var/reagents_per_slice = reagents.total_volume/slices_num
 	for(var/i=1 to (slices_num-slices_lost))
-		var/obj/slice = new slice_path (src.loc)
+		var/obj/item/weapon/reagent_containers/food/snacks/slice = new slice_path (loc)
+		initialize_slice(slice)
 		reagents.trans_to(slice,reagents_per_slice)
-	qdel(src) // so long and thanks for all the fish
+	qdel(src)
 
+/obj/item/weapon/reagent_containers/food/snacks/proc/initialize_slice(obj/item/weapon/reagent_containers/food/snacks/slice)
+	return
+
+/obj/item/weapon/reagent_containers/food/snacks/proc/update_overlays(obj/item/weapon/reagent_containers/food/snacks/S)
+	overlays.Cut()
+	var/image/I = new(src.icon, "[initial(icon_state)]_filling")
+	if(S.filling_color == "#FFFFFF")
+		I.color = pick("#FF0000","#0000FF","#008000","#FFFF00")
+	else
+		I.color = S.filling_color
+
+	overlays += I
+
+// cook() is called when microwaving the food
+/obj/item/weapon/reagent_containers/food/snacks/proc/initialize_cooked_food(obj/item/weapon/reagent_containers/food/snacks/S)
+	if(reagents)
+		reagents.trans_to(S, reagents.total_volume)
 
 /obj/item/weapon/reagent_containers/food/snacks/Destroy()
 	if(contents)
@@ -177,15 +214,14 @@
 			something.loc = get_turf(src)
 	..()
 
-
 /obj/item/weapon/reagent_containers/food/snacks/attack_animal(mob/M)
 	if(isanimal(M))
 		if(iscorgi(M))
 			if(bitecount == 0 || prob(50))
-				M.emote("me", 1, "nibbles away at the [src]")
+				M.emote("me", 1, "nibbles away at \the [src]")
 			bitecount++
 			if(bitecount >= 5)
-				var/sattisfaction_text = pick("burps from enjoyment", "yaps for more", "woofs twice", "looks at the area where the [src] was")
+				var/sattisfaction_text = pick("burps from enjoyment", "yaps for more", "woofs twice", "looks at the area where \the [src] was")
 				if(sattisfaction_text)
 					M.emote("me", 1, "[sattisfaction_text]")
 				qdel(src)
@@ -199,7 +235,7 @@
 
 //Notes by Darem: Food in the "snacks" subtype can hold a maximum of 50 units Generally speaking, you don't want to go over 40
 //	total for the item because you want to leave space for extra condiments. If you want effect besides healing, add a reagent for
-//	it. Try to stick to existing reagents when possible (so if you want a stronger healing effect, just use Tricordrazine). On use
+//	it. Try to stick to existing reagents when possible (so if you want a stronger healing effect, just use omnizine). On use
 //	effect (such as the old officer eating a donut code) requires a unique reagent (unless you can figure out a better way).
 
 //The nutriment reagent and bitesize variable replace the old heal_amt and amount variables. Each unit of nutriment is equal to
@@ -218,19 +254,19 @@
 //		reagents.add_reagent("nutriment", 2)							//	this line of code for all the contents.
 //		bitesize = 3													//This is the amount each bite consumes.
 
-//All foods (except slicable, since they have unique procs) are distributed among various categories. Use common sense.
+//All foods are distributed among various categories. Use common sense.
 
-/////////////////////////////////////////////////Sliceable////////////////////////////////////////
-// All the food items that can be sliced into smaller bits like Meatbread and Cheesewheels
+/////////////////////////////////////////////////Store////////////////////////////////////////
+// All the food items that can store an item inside itself, like bread or cake.
 
-//sliceable only changes w class, storage is handled by sliceable/store
-/obj/item/weapon/reagent_containers/food/snacks/sliceable
+
+/obj/item/weapon/reagent_containers/food/snacks/store
 	w_class = 3
 
-/obj/item/weapon/reagent_containers/food/snacks/sliceable/store
-
-/obj/item/weapon/reagent_containers/food/snacks/sliceable/store/attackby(obj/item/weapon/W, mob/user)
-	if(W.w_class <= 2)
+/obj/item/weapon/reagent_containers/food/snacks/store/attackby(obj/item/weapon/W, mob/user, params)
+	if(W.w_class > 2 || custom_food_type) //can't store objects inside food needed to start a customizable snack.
+		..()
+	else
 		if(contents.len)
 			return 0
 		if(!iscarbon(user))
@@ -240,29 +276,3 @@
 		add_fingerprint(user)
 		contents += W
 		return 1 // no afterattack here
-	else
-		..()
-
-/obj/item/weapon/reagent_containers/food/snacks/sliceable/cheesewheel
-	name = "cheese wheel"
-	desc = "A big wheel of delcious Cheddar."
-	icon_state = "cheesewheel"
-	slice_path = /obj/item/weapon/reagent_containers/food/snacks/cheesewedge
-	slices_num = 5
-
-/obj/item/weapon/reagent_containers/food/snacks/sliceable/cheesewheel/New()
-	..()
-	reagents.add_reagent("nutriment", 20)
-	bitesize = 2
-
-/obj/item/weapon/reagent_containers/food/snacks/cheesewedge
-	name = "cheese wedge"
-	desc = "A wedge of delicious Cheddar. The cheese wheel it was cut from can't have gone far."
-	icon_state = "cheesewedge"
-	bitesize = 2
-
-/obj/item/weapon/reagent_containers/food/snacks/watermelonslice
-	name = "watermelon slice"
-	desc = "A slice of watery goodness."
-	icon_state = "watermelonslice"
-	bitesize = 2
